refactor(reference): move the property reference to a separate module
This will make any interface convertible to a `PropertyRef` not only the
one returned by interface. Also this is now in the public interface so
it can be used to pass around a type-safe property reference.
